# internal/api/swagger/swagger.yaml
basePath: /v1
definitions:
...
  internal_api_handler.envelope:
    properties:
      data: {}
      error:
        type: string
    type: object
...
host: localhost:8000
info:
  contact:
    name: kith
    url: https://github.com/nhymxu/kith-pms
  description: Self-hosted personal relationship manager API.
  license:
    name: MIT
  title: kith API
  version: "1.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate with password. Sets kith_session HttpOnly cookie on
        success.
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          properties:
            password:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_api_handler.envelope'
            - properties:
                data:
                  properties:
                    logged_in:
                      type: boolean
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handler.envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api_handler.envelope'
      summary: Login
      tags:
      - auth
  /auth/logout:
    post:
      description: Revoke current session and clear the session cookie.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_api_handler.envelope'
            - properties:
                data:
                  properties:
                    logged_out:
                      type: boolean
                  type: object
              type: object
      security:
      - CookieAuth: []
      - CSRFHeader: []
      summary: Logout
      tags:
      - auth
  /auth/logout-all:
    post:
      description: Revoke all active sessions and clear the session cookie.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_api_handler.envelope'
            - properties:
                data:
                  properties:
                    logged_out:
                      type: boolean
                  type: object
              type: object
      security:
      - CookieAuth: []
      - CSRFHeader: []
      summary: Logout all sessions
      tags:
      - auth
  /auth/me:
    get:
      description: Returns the authenticated user's ID and creation time.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_api_handler.envelope'
            - properties:
                data:
                  properties:
                    created_at:
                      type: string
                    id:
                      type: integer
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api_handler.envelope'
      security:
      - CookieAuth: []
      summary: Get current user
      tags:
      - auth
  /auth/password:
    post:
      consumes:
      - application/json
      description: Change the login password. Revokes all sessions and re-issues a
        new one.
      parameters:
      - description: Password change request
        in: body
        name: body
        required: true
        schema:
          properties:
            confirm_password:
              type: string
            current_password:
              type: string
            new_password:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_api_handler.envelope'
            - properties:
                data:
                  properties:
                    password_changed:
                      type: boolean
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handler.envelope'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/internal_api_handler.envelope'
      security:
      - CookieAuth: []
      - CSRFHeader: []
      summary: Change password
      tags:
      - auth
...
securityDefinitions:
  CSRFHeader:
    description: Must be set to "kith-spa" on all mutating requests (POST/PUT/PATCH/DELETE)
    in: header
    name: X-Requested-With
    type: apiKey
  CookieAuth:
    description: HttpOnly session cookie set by POST /v1/auth/login
    in: cookie
    name: kith_session
    type: apiKey
swagger: "2.0"